requirements.txt
matplotlib
mysql-connector
mysql-connector-python
python-dotenv